# Reduce allocations and memory usage in the loggers