This PR refactors the loggers to reduce the number of allocations and amount of memory needed for each log line. Benchmarks are done by assuming 0, 3, 9, 15, and 30 total arguments.

Before:
```
nacx ~/src/log (v2) $ go test -bench=.

BenchmarkStructuredLog0Args-8      	 1215139	       925.9 ns/op	     328 B/op	       9 allocs/op
BenchmarkStructuredLog3Args-8      	  825308	      1379 ns/op	     624 B/op	      18 allocs/op
BenchmarkStructuredLog9Args-8      	  562828	      2076 ns/op	    1216 B/op	      30 allocs/op
BenchmarkStructuredLog15Args-8     	  412002	      2846 ns/op	    2096 B/op	      43 allocs/op
BenchmarkStructuredLog30Args-8     	  259807	      4813 ns/op	    3577 B/op	      73 allocs/op

BenchmarkUnstructuredLog0Args-8    	 1735750	       688.6 ns/op	     144 B/op	       5 allocs/op
BenchmarkUnstructuredLog3Args-8    	  989944	      1152 ns/op	     536 B/op	      13 allocs/op
BenchmarkUnstructuredLog9Args-8    	  651796	      1794 ns/op	     952 B/op	      21 allocs/op
BenchmarkUnstructuredLog15Args-8   	  491506	      2402 ns/op	    1400 B/op	      29 allocs/op
BenchmarkUnstructuredLog30Args-8   	  295832	      4042 ns/op	    3177 B/op	      50 allocs/op
```

After:
```
nacx ~/src/log (format) $ go test -bench=.

BenchmarkStructuredLog0Args-8      	 1453257	       776.0 ns/op	     312 B/op	       6 allocs/op
BenchmarkStructuredLog3Args-8      	 1227420	       987.8 ns/op	     312 B/op	       6 allocs/op
BenchmarkStructuredLog9Args-8      	  792440	      1515 ns/op	     600 B/op	       7 allocs/op
BenchmarkStructuredLog15Args-8     	  611298	      2145 ns/op	     600 B/op	       7 allocs/op
BenchmarkStructuredLog30Args-8     	  413978	      2881 ns/op	    1208 B/op	       8 allocs/op

BenchmarkUnstructuredLog0Args-8    	 1811126	       647.8 ns/op	     144 B/op	       5 allocs/op
BenchmarkUnstructuredLog3Args-8    	 1282764	       988.5 ns/op	     360 B/op	       6 allocs/op
BenchmarkUnstructuredLog9Args-8    	  843518	      1411 ns/op	     424 B/op	       6 allocs/op
BenchmarkUnstructuredLog15Args-8   	  668487	      1747 ns/op	     520 B/op	       6 allocs/op
BenchmarkUnstructuredLog30Args-8   	  425914	      2847 ns/op	    1416 B/op	       7 allocs/op
```