@nacx 2021-12-11
@basvanbeek you may want to incorporate some of this into the go-kit implementation, as the initial version of this logger was a copy of that one.